#152 - refactoring: renamed classes
