+++ src/main/java/org/opengis/cite/ogcapifeatures10/conformance/crs/query/crs/feature/FeatureCrsParameterTransform.java
@@ -38,7 +38,7 @@ import io.restassured.response.Response;
  *
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
  */
-public class FeatureValidCrsParameterTransform extends AbstractFeatureCrs {
+public class FeatureCrsParameterTransform extends AbstractFeatureCrs {
 
     /**
      * Test: invalid CRS requesting /collections/{collectionId}/items
