+++ src/main/java/org/opengis/cite/ogcapifeatures10/conformance/crs/query/crs/features/FeaturesCrsParameter.java
@@ -1,10 +1,10 @@
-package org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.feature;
+package org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.features;
 
 import static io.restassured.http.Method.GET;
 import static org.opengis.cite.ogcapifeatures10.EtsAssert.assertCrsHeader;
 import static org.opengis.cite.ogcapifeatures10.OgcApiFeatures10.CRS_PARAMETER;
 import static org.opengis.cite.ogcapifeatures10.OgcApiFeatures10.GEOJSON_MIME_TYPE;
-import static org.opengis.cite.ogcapifeatures10.util.JsonUtils.findFeatureUrlForGeoJson;
+import static org.opengis.cite.ogcapifeatures10.util.JsonUtils.findFeaturesUrlForGeoJson;
 
 import org.testng.SkipException;
 import org.testng.annotations.Test;
@@ -13,8 +13,8 @@ import io.restassured.path.json.JsonPath;
 import io.restassured.response.Response;
 
 /**
- * Verifies header in the paths /collection/{collectionId}/items/{featureId}
- *
+ * Verifies header in the paths /collection/{collectionId}/items
+ * 
  * <pre>
  * Abstract Test 4: /conf/crs/crs-parameter
  * Test Purpose: Verify that the parameter crs has been implemented correctly
@@ -33,42 +33,40 @@ import io.restassured.response.Response;
  *  * has the status code 200 and
  *  * includes a Content-Crs http header with the value of the requested CRS identifier.
  * </pre>
- *
+ * 
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
  */
-public class FeatureValidCrsParameter extends AbstractFeatureCrs {
+public class FeaturesCrsParameter extends AbstractFeaturesCrs {
 
     /**
-     * Test: Content-Crs header in the path /collections/{collectionId}/items/{featureId}
+     * Test: Content-Crs header in the path /collections/{collectionId}/items
      *
      * @param collectionId
      *            id id of the collection, never <code>null</code>
      * @param collection
      *            the /collection object, never <code>null</code>
-     * @param featureId
-     *            id id of the feature, never <code>null</code>
      * @param crs
      *            the crs to test, never <code>null</code>
      */
     @Test(description = "Implements A.2.1 Query, Parameter crs, Abstract Test 1 (Requirement /req/crs/fc-crs-definition, /req/crs/fc-crs-valid-value B, /req/crs/ogc-crs-header, /req/crs/ogc-crs-header-value, /req/crs/geojson), "
-                        + "Content-Crs header in the path /collections/{collectionId}/items/{featureId}", dataProvider = "collectionFeatureIdCrs", dependsOnGroups = "crs-conformance")
-    public void verifyFeaturePathCrsHeader( String collectionId, JsonPath collection, String featureId, String crs ) {
-        String featureUrl = findFeatureUrlForGeoJson( rootUri, collection, featureId );
-        if ( featureUrl == null )
+                        + "Content-Crs header in the path /collections/{collectionId}/items", dataProvider = "collectionIdAndJsonAndCrs", dependsOnGroups = "crs-conformance")
+    public void verifyFeaturesPathCrsHeader( String collectionId, JsonPath collection, String crs ) {
+        String featuresUrl = findFeaturesUrlForGeoJson( rootUri, collection );
+        if ( featuresUrl == null )
             throw new SkipException( "Could not find url for collection with id " + collectionId
                                      + " supporting GeoJson (type " + GEOJSON_MIME_TYPE + ")" );
 
-        Response response = init().baseUri( featureUrl ).queryParam( CRS_PARAMETER,
-                                                                     crs ).accept( GEOJSON_MIME_TYPE ).when().request( GET );
+        Response response = init().baseUri( featuresUrl ).queryParam( CRS_PARAMETER,
+                                                                      crs ).accept( GEOJSON_MIME_TYPE ).when().request( GET );
         response.then().statusCode( 200 );
         String actualHeader = response.getHeader( "Content-Crs" );
         if ( actualHeader == null ) {
-            throw new AssertionError( String.format( "Feature response at '%s' does not provide the expected header 'Content-Crs'",
-                                                     featureUrl ) );
+            throw new AssertionError( String.format( "Features response at '%s' does not provide the expected header 'Content-Crs'",
+                                                     featuresUrl ) );
         }
         assertCrsHeader( actualHeader, crs,
                          String.format( "Features response at '%s' does not provide expected 'Content-Crs' header, was: '%s', expected: '%s'",
-                                        featureUrl, actualHeader, crs ) );
+                                        featuresUrl, actualHeader, crs ) );
     }
 
 }
